refactored kotlin activities to make project navigation easier
