+++ app/src/main/AndroidManifest.xml
@@ -12,30 +12,21 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SignifyBasic"
         tools:targetApi="31" >
-        <activity
-            android:name=".test_knowledge"
-            android:exported="false"
-            android:label="@string/title_activity_test_knowledge"
-            android:theme="@style/Theme.SignifyBasic" />
-        <activity
-            android:name=".ProfileScreen"
-            android:exported="true" />
-        <activity android:name=".SignupPage" />
-        <activity android:name=".WelcomeCenter" />
-        <activity android:name=".TestPage" />
+        <activity android:name=".features.auth.SignupPage" />
+        <activity android:name=".features.tabs.HomePage" />
+        <activity android:name=".features.auth.TestPage" />
         <activity android:name=".MainActivity3" />
         <activity android:name=".TakePictureActivity" />
         <activity android:name=".LoadImageActivity" />
-        <activity android:name=".HomeActivity" />
-        <activity android:name=".AchievementsActivity" />
-        <activity android:name=".ProfileActivity" />
-        <activity android:name=".NotificationsActivity" />
-        <activity android:name=".SettingsActivity" />
-        <activity android:name=".DiscussionActivity" />
-        <activity android:name=".ActivityCenterActivity" />
-        <activity android:name=".ResourcesActivity" />
+        <activity android:name=".features.tabs.AchievementsPage" />
+        <activity android:name=".features.tabs.ProfilePage" />
+        <activity android:name=".features.tabs.NotificationsPage" />
+        <activity android:name=".features.tabs.SettingsPage" />
+        <activity android:name=".features.tabs.DiscussionPage" />
+        <activity android:name=".features.activitycenter.ActivityCenter" />
+        <activity android:name=".features.tabs.ResourcesPage" />
         <activity
-            android:name=".MainActivity"
+            android:name=".features.auth.MainActivity"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
