--- app/src/main/java/com/example/signifybasic/DiscussionActivity.kt
@@ -1,11 +1,0 @@
-package com.example.signifybasic
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class DiscussionActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_discussion)
-    }
-}
