+++ app/src/main/java/com/example/signifybasic/features/activitycenter/ActivityCenter.kt
@@ -1,9 +1,10 @@
-package com.example.signifybasic
+package com.example.signifybasic.features.activitycenter
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.signifybasic.R
 
-class ActivityCenterActivity : AppCompatActivity() {
+class ActivityCenter : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_activity_center)
