+++ app/src/main/java/com/example/signifybasic/features/auth/SignupPage.kt
@@ -1,33 +1,34 @@
-package com.example.signifybasic
+package com.example.signifybasic.features.auth
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
-import android.widget.EditText
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.example.signifybasic.R
 
 
-class TestPage : AppCompatActivity() {
+class SignupPage : AppCompatActivity() {
 
-    private lateinit var continueButton : Button
+    lateinit var signupButton : Button
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.test_page)
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.test)) { v, insets ->
+        setContentView(R.layout.signup_page)
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.signup)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
         }
 
-        // temporary implementation : just continue to welcome center.
-        continueButton = findViewById(R.id.continue_button)
-        continueButton.setOnClickListener {
-            val intent = Intent(this, WelcomeCenter::class.java)
+        signupButton = findViewById(R.id.signup_button)
+
+        // once user signs up, they should be routed to the 'test' page
+        signupButton.setOnClickListener {
+            val intent = Intent(this, TestPage::class.java)
             startActivity(intent)
         }
     }
