+++ app/src/main/java/com/example/signifybasic/features/auth/TestPage.kt
@@ -1,4 +1,4 @@
-package com.example.signifybasic
+package com.example.signifybasic.features.auth
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,27 +7,28 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.example.signifybasic.R
+import com.example.signifybasic.features.tabs.HomePage
 
 
-class SignupPage : AppCompatActivity() {
+class TestPage : AppCompatActivity() {
 
-    lateinit var signupButton : Button
+    private lateinit var continueButton : Button
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.signup_page)
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.signup)) { v, insets ->
+        setContentView(R.layout.test_page)
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.test)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
         }
 
-        signupButton = findViewById(R.id.signup_button)
-
-        // once user signs up, they should be routed to the 'test' page
-        signupButton.setOnClickListener {
-            val intent = Intent(this, TestPage::class.java)
+        // temporary implementation : just continue to welcome center.
+        continueButton = findViewById(R.id.continue_button)
+        continueButton.setOnClickListener {
+            val intent = Intent(this, HomePage::class.java)
             startActivity(intent)
         }
     }
