+++ app/src/main/java/com/example/signifybasic/features/tabs/AchievementsPage.kt
@@ -1,9 +1,10 @@
-package com.example.signifybasic
+package com.example.signifybasic.features.tabs
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.signifybasic.R
 
-class AchievementsActivity : AppCompatActivity() {
+class AchievementsPage : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_achievements)
