+++ app/src/main/java/com/example/signifybasic/features/tabs/DiscussionPage.kt
@@ -1,11 +1,12 @@
-package com.example.signifybasic
+package com.example.signifybasic.features.tabs
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.signifybasic.R
 
-class HomeActivity : AppCompatActivity() {
+class DiscussionPage : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_home)
+        setContentView(R.layout.activity_discussion)
     }
 }
