+++ app/src/main/res/layout/activity_load_image.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".features.auth.MainActivity">
 
     <!-- add Camera Button to open the Camera -->
     <Button
