+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:background="@drawable/login_background"
     android:gravity="center"
     android:padding="16dp"
-    tools:context=".MainActivity">
+    tools:context=".features.auth.MainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
