+++ app/src/main/res/layout/activity_profile.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/login_background"
-    tools:context=".MainActivity">
+    tools:context=".features.auth.MainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
