+++ app/src/main/res/layout/notifications_activity.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".TestPage">
+    tools:context=".features.auth.TestPage">
 
     <LinearLayout
         android:layout_width="match_parent"
