+++ app/src/main/res/layout/signup_page.xml
@@ -8,7 +8,7 @@
     android:background="@drawable/login_background"
     android:gravity="center"
     android:padding="16dp"
-    tools:context=".SignupPage">
+    tools:context=".features.auth.SignupPage">
 
     <LinearLayout
         android:layout_width="match_parent"
