+++ app/src/main/res/layout/test_page.xml
@@ -8,7 +8,7 @@
     android:background="@drawable/login_background"
     android:gravity="center"
     android:padding="16dp"
-    tools:context=".TestPage">
+    tools:context=".features.auth.TestPage">
 
     <LinearLayout
         android:layout_width="match_parent"
