+++ app/src/main/res/layout/welcome_center.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WelcomeCenter">
+    tools:context=".features.tabs.HomePage">
 
     <View
         android:layout_width="match_parent"
